requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
boto3
google-analytics-data==0.18.18
google-api-core==2.24.2
google-auth==2.40.1
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
idna==3.10
proto-plus==1.26.1
protobuf==6.31.0rc2
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
urllib3==2.4.0
pyiceberg
pyiceberg[s3fs,pyarrow]
pyarrow